Combined cPathFinder and cPath
